removed unnecessary trait
