+++ manta-pay/src/signer/client/websocket.rs
@@ -42,16 +42,6 @@ use tokio_tungstenite::{
     MaybeTlsStream, WebSocketStream,
 };
 
-#[cfg(feature = "wallet")]
-use {
-    crate::signer::stateless_signer::{
-        StatelessAddressRequest, StatelessIdentityRequest, StatelessSignRequest,
-        StatelessSignResult, StatelessSyncRequest, StatelessSyncResult,
-        StatelessTransactionDataRequest,
-    },
-    manta_accounting::wallet::stateless_signer,
-};
-
 /// Web Socket Error
 pub type WebSocketError = tungstenite::error::Error;
 
@@ -175,51 +165,6 @@ impl signer::Connection<Config> for Client {
     fn identity_proof(
         &mut self,
         request: IdentityRequest,
-    ) -> LocalBoxFutureResult<IdentityResponse, Self::Error> {
-        Box::pin(async move { self.send("identity", request).await })
-    }
-}
-
-#[cfg(feature = "wallet")]
-impl stateless_signer::StatelessSignerConnection<Config> for Client {
-    type Error = Error;
-
-    #[inline]
-    fn sync(
-        &mut self,
-        request: StatelessSyncRequest,
-    ) -> LocalBoxFutureResult<StatelessSyncResult, Self::Error> {
-        Box::pin(async move { self.send("sync", request).await })
-    }
-
-    #[inline]
-    fn sign(
-        &mut self,
-        request: StatelessSignRequest,
-    ) -> LocalBoxFutureResult<StatelessSignResult, Self::Error> {
-        Box::pin(async move { self.send("sign", request).await })
-    }
-
-    #[inline]
-    fn address(
-        &mut self,
-        request: StatelessAddressRequest,
-    ) -> LocalBoxFutureResult<Address, Self::Error> {
-        Box::pin(async move { self.send("address", request).await })
-    }
-
-    #[inline]
-    fn transaction_data(
-        &mut self,
-        request: StatelessTransactionDataRequest,
-    ) -> LocalBoxFutureResult<TransactionDataResponse, Self::Error> {
-        Box::pin(async move { self.send("transaction_data", request).await })
-    }
-
-    #[inline]
-    fn identity_proof(
-        &mut self,
-        request: StatelessIdentityRequest,
     ) -> LocalBoxFutureResult<IdentityResponse, Self::Error> {
         Box::pin(async move { self.send("identity", request).await })
     }
